Add files via upload
Uploaded a PlatformIO project as a zip file, which contains the code used to test. Pin mappings are included in comments at the top of the corresponding cpp file. I believe you will need to rename the file to "main.cpp" to run it. However, for now, they are named "tof.cpp" and "ultrasonic.cpp".